requirements.txt
git+https://github.com/hakutakaid/pyrogram
tgcrypto
pyromod==1.5
python-dotenv